+++ Runtime/Scripts/GltfImport.cs
@@ -1725,7 +1725,7 @@ namespace GLTFast {
                     // await defaultDeferAgent.BreakPoint();
 
                     if(primitive.HasValue) {
-                        primitives[primitiveContext.primtiveIndex] = primitive.Value;
+                        primitives[primitiveContext.primitiveIndex] = primitive.Value;
                         resources.Add(primitive.Value.mesh);
                     } else {
                         success = false;
@@ -2776,19 +2776,24 @@ namespace GLTFast {
                             if (!bounds.HasValue) {
                                 logger.Error(LogCode.MeshBoundsMissing, meshIndex.ToString());
                             }
-                            context = new PrimitiveDracoCreateContext(bounds);
-                            context.materials = new int[1];
+                            var dracoContext = new PrimitiveDracoCreateContext(
+                                primitiveIndex,
+                                1,
+                                primitive.material<0 || gltf.materials[primitive.material].requiresNormals,
+                                primitive.material>=0 && gltf.materials[primitive.material].requiresTangents,
+                                mesh.name,
+                                bounds
+                                );
+                            context = dracoContext;
                         }
                         else
 #endif
                         {
                             PrimitiveCreateContext c;
                             if(context==null) {
-                                c = new PrimitiveCreateContext();
-                                c.indices = new int[cluster.Count][];
-                                c.materials = new int[cluster.Count];
+                                c = new PrimitiveCreateContext(primitiveIndex, cluster.Count, mesh.name);
                             } else {
-                                c = (context as PrimitiveCreateContext);
+                                c = context as PrimitiveCreateContext;
                             }
                             // PreparePrimitiveIndices(gltf,mesh,primitive,ref c,primIndex);
                             context = c;
@@ -2798,11 +2803,7 @@ namespace GLTFast {
                             context.morphTargetsContext = morphTargetsContexts[primitive];
                         }
                         
-                        context.primtiveIndex = primitiveIndex;
-                        context.materials[primIndex] = primitive.material;
-
-                        context.needsNormals |= primitive.material<0 || gltf.materials[primitive.material].requiresNormals;
-                        context.needsTangents |= primitive.material>=0 && gltf.materials[primitive.material].requiresTangents;
+                        context.SetMaterial(primIndex, primitive.material);
                     }
 
                     primitiveContexts[primitiveIndex] = context;
@@ -2891,27 +2892,6 @@ namespace GLTFast {
         }
 
         async Task AssignAllAccessorData( Root gltf ) {
-            if (gltf.meshes != null) {
-                Profiler.BeginSample("AssignAllAccessorData.Primitive");
-                int i=0;
-                for( int meshIndex = 0; meshIndex<gltf.meshes.Length; meshIndex++ ) {
-                    var mesh = gltf.meshes[meshIndex];
-
-                    foreach( var cluster in meshPrimitiveCluster[meshIndex]) {
-    #if DRACO_UNITY
-                        if( !cluster.Value[0].isDracoCompressed )
-    #endif
-                        {
-                            // Create one PrimitiveCreateContext per Primitive cluster
-                            PrimitiveCreateContext c = (PrimitiveCreateContext) primitiveContexts[i];
-                            c.mesh = mesh;
-                        }
-                        i++;
-                    }
-                }
-                Profiler.EndSample();
-            }
-            
             if(gltf.skins!=null) {
                 for (int s = 0; s < gltf.skins.Length; s++)
                 {
@@ -2926,7 +2906,7 @@ namespace GLTFast {
             }
         }
 
-        void PreparePrimitiveIndices( Root gltf, Mesh mesh, MeshPrimitive primitive, ref PrimitiveCreateContext c, int submeshIndex = 0 ) {
+        void PreparePrimitiveIndices( Root gltf, Mesh mesh, MeshPrimitive primitive, ref PrimitiveCreateContext c, int subMesh = 0 ) {
             Profiler.BeginSample("PreparePrimitiveIndices");
             switch(primitive.mode) {
             case DrawMode.Triangles:
@@ -2954,11 +2934,12 @@ namespace GLTFast {
             }
 
             if(primitive.indices >= 0) {
-                c.indices[submeshIndex] = (accessorData[primitive.indices] as AccessorData<int>).data;
+                c.SetIndices(subMesh, ((AccessorData<int>)accessorData[primitive.indices]).data);
             } else {
                 int vertexCount = gltf.accessors[primitive.attributes.POSITION].count;
                 JobHandle? jh;
-                CalculateIndicesJob(gltf,primitive, vertexCount, c.topology, out c.indices[submeshIndex], out jh, out c.calculatedIndicesHandle );
+                CalculateIndicesJob(gltf, primitive, vertexCount, c.topology, out var indices, out jh, out c.calculatedIndicesHandle);
+                c.SetIndices(subMesh, indices);
                 c.jobHandle = jh.Value;
             }
             Profiler.EndSample();