+++ Runtime/Scripts/PrimitiveCreateContextBase.cs
@@ -18,8 +18,6 @@ using System.Threading.Tasks;
 
 namespace GLTFast {
     
-    using Schema;
-    
     abstract class PrimitiveCreateContextBase {
         
         public const MeshUpdateFlags defaultMeshUpdateFlags =
@@ -27,12 +25,23 @@ namespace GLTFast {
             | MeshUpdateFlags.DontRecalculateBounds
             | MeshUpdateFlags.DontResetBoneBounds
             | MeshUpdateFlags.DontValidateIndices;
-        
-        public int primtiveIndex;
-        public int[] materials;
-        public bool needsNormals;
-        public bool needsTangents;
+
+        protected string m_MeshName;
+        protected int[] m_Materials;
+
+        public int primitiveIndex { get; }
+
         public abstract bool IsCompleted {get;}
+
+        protected PrimitiveCreateContextBase(int primitiveIndex, int materialCount, string meshName) {
+            this.primitiveIndex = primitiveIndex;
+            m_Materials = new int[materialCount];
+            m_MeshName = meshName;
+        }
+
+        public void SetMaterial(int subMesh, int materialIndex) {
+            m_Materials[subMesh] = materialIndex;
+        }
         
         public MorphTargetsContext morphTargetsContext;